+++ t/app/controller/report_new.t
@@ -159,7 +159,7 @@ foreach my $test (
                     username_register => 'test-1@example.com',
                     phone         => '07903 123 456',
                     category      => 'Street lighting',
-                    password_register => $test->{password} ? 'secret' : '',
+                    password_register => $test->{password} ? 'secretsecret' : '',
                 }
             },
             "submit good details"
@@ -208,7 +208,7 @@ foreach my $test (
 
     is $report->name, 'Joe Bloggs', 'name updated correctly';
     if ($test->{password}) {
-        ok $report->user->check_password('secret'), 'password updated correctly';
+        ok $report->user->check_password('secretsecret'), 'password updated correctly';
     } elsif ($test->{user}) {
         ok $report->user->check_password('old_password'), 'password unchanged, as no new one given';
     } else {